// src/shared/api/generated/schemas/auctionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Статус аукциона:
 * - `Planning` — планирование (1)
 * - `Auction` — торги идут (2)
 * - `DeterminateWinner` — определение победителя (3)
 * - `WaitDeal` — ожидание сделки (4)
 * - `InProgress` — в работе (5)
 * - `Finished` — завершён (6)
 * - `Stopped` — остановлен (7)
 * - `Canceled` — отменён (8)
 * - `Unknown` — неизвестный статус
 */
export type AuctionStatus = typeof AuctionStatus[keyof typeof AuctionStatus];


export const AuctionStatus = {
  Planning: 'Planning',
  Auction: 'Auction',
  DeterminateWinner: 'DeterminateWinner',
  WaitDeal: 'WaitDeal',
  InProgress: 'InProgress',
  Finished: 'Finished',
  Stopped: 'Stopped',
  Canceled: 'Canceled',
  Unknown: 'Unknown',
} as const;