// src/shared/api/generated/schemas/bidMeasurementType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * API для UL - Auctions
 * API для мобильных приложений и web UL
 * OpenAPI spec version: 1.0.0
 */

/**
 * Единица измерения ставки:
 * - `PerRoute` — за рейс (0)
 * - `PerKm` — за км (1)
 * - `Unknown` — неизвестный тип
 */
export type BidMeasurementType = typeof BidMeasurementType[keyof typeof BidMeasurementType];


export const BidMeasurementType = {
  PerRoute: 'PerRoute',
  PerKm: 'PerKm',
  Unknown: 'Unknown',
} as const;